+++ app/Http/Controllers/Api/CertificationController.php
@@ -1,18 +1,20 @@
 <?php
 
 namespace App\Http\Controllers\Api;
+use App\Helpers\GeneralHelpers;
 use App\Models\Certification;
 use App\Models\CertificationElement;
 use App\Models\Module;
 use App\Models\TrainingUnit;
 use Illuminate\Http\Request;
-use App\Services\TrainingContractService;
+use Illuminate\Support\Facades\Auth;
 
 class CertificationController extends BaseController
 {
-    public function certifications() {
+    public function certifications(Request $request) {
         try {
-            return Certification::getCertifications();
+           $mainCompanyId = GeneralHelpers::urlObtainCompanyId($request->headers->get('origin'), Auth::id());
+            return Certification::getCertifications($mainCompanyId);
         } catch (\Exception $e) {
             return response()->json([
                 'message' => $e->getMessage()
@@ -22,7 +24,11 @@ class CertificationController extends BaseController
 
     public function create(Request $request){
         try {
-            $certification = Certification::createCertification($request);
+           $mainCompanyId = GeneralHelpers::urlObtainCompanyId($request->headers->get('origin'), Auth::id());
+            $data = $request->all();
+            $data['main_company_id'] = $mainCompanyId;
+
+            $certification = Certification::createCertification($data);
         } catch (\Exception $e){
             return response()->json([
                 'status' => 400,
@@ -32,13 +38,17 @@ class CertificationController extends BaseController
 
         return response()->json([
             'status' => 200,
-            'certification' => Certification::getCertification($certification->id)
+            'certification' => Certification::getCertification($certification->id, $mainCompanyId)
         ]);
     }
 
     public function edit($id, Request $request){
         try {
-            $certification = Certification::updateCertification($id, $request);
+           $mainCompanyId = GeneralHelpers::urlObtainCompanyId($request->headers->get('origin'), Auth::id());
+            $data = $request->all();
+            $data['main_company_id'] = $mainCompanyId;
+
+            $certification = Certification::updateCertification($id, $data);
         } catch (\Exception $e){
             return response()->json([
                 'status' => 400,
@@ -48,12 +58,14 @@ class CertificationController extends BaseController
 
         return response()->json([
             'status' => 200,
-            'certification' => Certification::getCertification($certification->id)
+            'certification' => Certification::getCertification($certification->id, $mainCompanyId)
         ]);
     }
 
-    public function getCertification($id){
-        $certification = Certification::getCertification($id);
+    public function getCertification($id, Request $request){
+       $mainCompanyId = GeneralHelpers::urlObtainCompanyId($request->headers->get('origin'), Auth::id());
+        $certification = Certification::getCertification($id, $mainCompanyId);
+
         if ($certification) {
             return response()->json([
                 'status' => 200,
@@ -66,9 +78,22 @@ class CertificationController extends BaseController
         ]);
     }
 
-    public function destroy($id){
+    public function destroy($id, Request $request){
         if ($id) {
             try {
+               $mainCompanyId = GeneralHelpers::urlObtainCompanyId($request->headers->get('origin'), Auth::id());
+
+                $certification = Certification::where('id', $id)
+                    ->FilterMainCompany($mainCompanyId)
+                    ->first();
+
+                if (!$certification) {
+                    return response()->json([
+                        'status' => 404,
+                        'message' => 'Certificado no existe'
+                    ]);
+                }
+
                 Certification::destroy($id);
                 return response()->json([
                     'status' => 200
@@ -86,13 +111,18 @@ class CertificationController extends BaseController
      * Volver a calcular horas de los certificados
      * @return void
      */
-    public function recalculateHours(){
-        $certifications = Certification::all();
+    public function recalculateHours(Request $request){
+       $mainCompanyId = GeneralHelpers::urlObtainCompanyId($request->headers->get('origin'), Auth::id());
+
+        $certifications = Certification::FilterMainCompany($mainCompanyId)->get();
+
         foreach($certifications as $certification) {
-            $hours = 0;
             $face_to_face_hours = 0;
             $teletraining_hours = 0;
-            $certification_elements = CertificationElement::where('certification_id', $certification->id)->get();
+            $certification_elements = CertificationElement::where('certification_id', $certification->id)
+                ->FilterMainCompany($mainCompanyId)
+                ->get();
+
             foreach ($certification_elements as $certification_element) {
                 if ($certification_element){
                     if ($certification_element->training_unit_id) {
