+++ app/Http/Controllers/Api/ChoreController.php
@@ -1,25 +1,15 @@
 <?php
 
 namespace App\Http\Controllers\Api;
+use App\Helpers\GeneralHelpers;
 use App\Models\Chore;
-use App\Models\Course;
-use App\Models\Student;
 use App\Models\User;
-use App\Services\ChoreService;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Facades\Validator;
 
 class ChoreController extends BaseController
 {
-    private $choreService;
-
-    public function __construct(ChoreService $choreService)
-    {
-        $this->choreService = $choreService;
-    }
 
     /**
      * Obtenemos tareas
@@ -27,13 +17,13 @@ class ChoreController extends BaseController
      */
     public function index(Request $request) {
         try {
+           $mainCompanyId = GeneralHelpers::urlObtainCompanyId($request->headers->get('origin'), Auth::id());
             $start = \Illuminate\Support\Carbon::now();
             $number_days = 3;
             if ($start->dayOfWeek >= 3)
                 $number_days = 5;
-            $start = $start->addDays($number_days);
 
-            $chores = Chore::chore();
+            $chores = Chore::chore($mainCompanyId);
             $user = User::find(Auth::id());
             if ($user->teacher_id) {
                 $chores = $chores->where('courses.teacher_id', $user->teacher_id);
@@ -173,8 +163,9 @@ class ChoreController extends BaseController
      * @param $id
      * @return \Illuminate\Http\JsonResponse
      */
-    public function show($id){
-        $chore = Chore::chore()
+    public function show($id, Request $request) {
+       $mainCompanyId = GeneralHelpers::urlObtainCompanyId($request->headers->get('origin'), Auth::id());
+        $chore = Chore::chore($mainCompanyId)
             ->where('chores.id', $id)
             ->first();
         if ($chore) {
@@ -289,93 +280,102 @@ class ChoreController extends BaseController
      */
     public function update($id, Request $request){
         try {
-            $chore = Chore::find($id);
-            $data = $request->all();
-            $element = $this->choreService->update($chore, $data);
-            $chore = Chore::chore()
-                ->where('chores.id', $element->id)
+           $mainCompanyId = GeneralHelpers::urlObtainCompanyId($request->headers->get('origin'), Auth::id());
+            $chore = Chore::where('id', $id)
+                ->FilterMainCompany($mainCompanyId)
                 ->first();
             if ($chore) {
-                if ($chore->membership_tab_status == 0) {
-                    $chore['membership_tab_status_name'] = 'Pendiente';
-                } else if ($chore->membership_tab_status == 1) {
-                    $chore['membership_tab_status_name'] = 'Enviada';
-                } else if ($chore->membership_tab_status == 2) {
-                    $chore['membership_tab_status_name'] = 'Recibida';
-                } else if ($chore->membership_tab_status == 3) {
-                    $chore['membership_tab_status_name'] = 'No procede';
-                }
+                $data = $request->all();
+                $element = $chore->updateWithService($data);
+                $chore = Chore::chore($mainCompanyId)
+                    ->where('chores.id', $element->id)
+                    ->first();
+                if ($chore) {
+                    if ($chore->membership_tab_status == 0) {
+                        $chore['membership_tab_status_name'] = 'Pendiente';
+                    } else if ($chore->membership_tab_status == 1) {
+                        $chore['membership_tab_status_name'] = 'Enviada';
+                    } else if ($chore->membership_tab_status == 2) {
+                        $chore['membership_tab_status_name'] = 'Recibida';
+                    } else if ($chore->membership_tab_status == 3) {
+                        $chore['membership_tab_status_name'] = 'No procede';
+                    }
 
-                if ($chore->economic_proposal_status == 0) {
-                    $chore['economic_proposal_status_name'] = 'Pendiente';
-                } else if ($chore->economic_proposal_status == 1) {
-                    $chore['economic_proposal_status_name'] = 'Enviada';
-                } else if ($chore->economic_proposal_status == 2) {
-                    $chore['economic_proposal_status_name'] = 'Recibida';
-                }
+                    if ($chore->economic_proposal_status == 0) {
+                        $chore['economic_proposal_status_name'] = 'Pendiente';
+                    } else if ($chore->economic_proposal_status == 1) {
+                        $chore['economic_proposal_status_name'] = 'Enviada';
+                    } else if ($chore->economic_proposal_status == 2) {
+                        $chore['economic_proposal_status_name'] = 'Recibida';
+                    }
 
-                if ($chore->student_tab_status == 0) {
-                    $chore['student_tab_status_name'] = 'Pendiente';
-                } else if ($chore->student_tab_status == 1) {
-                    $chore['student_tab_status_name'] = 'Enviada';
-                } else if ($chore->student_tab_status == 2) {
-                    $chore['student_tab_status_name'] = 'Recibida';
-                }
+                    if ($chore->student_tab_status == 0) {
+                        $chore['student_tab_status_name'] = 'Pendiente';
+                    } else if ($chore->student_tab_status == 1) {
+                        $chore['student_tab_status_name'] = 'Enviada';
+                    } else if ($chore->student_tab_status == 2) {
+                        $chore['student_tab_status_name'] = 'Recibida';
+                    }
 
-                if ($chore->welcome_guid_status == 0) {
-                    $chore['welcome_guid_status_name'] = 'Pendiente';
-                } else if ($chore->welcome_guid_status == 1) {
-                    $chore['welcome_guid_status_name'] = 'Realizada';
-                }
+                    if ($chore->welcome_guid_status == 0) {
+                        $chore['welcome_guid_status_name'] = 'Pendiente';
+                    } else if ($chore->welcome_guid_status == 1) {
+                        $chore['welcome_guid_status_name'] = 'Realizada';
+                    }
 
-                if ($chore->registration_status == 0) {
-                    $chore['registration_status_name'] = 'Pendiente';
-                } else if ($chore->registration_status == 1) {
-                    $chore['registration_status_name'] = 'Realizada';
-                }
+                    if ($chore->registration_status == 0) {
+                        $chore['registration_status_name'] = 'Pendiente';
+                    } else if ($chore->registration_status == 1) {
+                        $chore['registration_status_name'] = 'Realizada';
+                    }
 
-                if ($chore->diploma_status == 0) {
-                    $chore['diploma_status_name'] = 'Pendiente';
-                } else if ($chore->diploma_status == 1) {
-                    $chore['diploma_status_name'] = 'Realizada';
-                } else if ($chore->diploma_status == 2) {
-                    $chore['diploma_status_name'] = 'No procede';
-                }
+                    if ($chore->diploma_status == 0) {
+                        $chore['diploma_status_name'] = 'Pendiente';
+                    } else if ($chore->diploma_status == 1) {
+                        $chore['diploma_status_name'] = 'Realizada';
+                    } else if ($chore->diploma_status == 2) {
+                        $chore['diploma_status_name'] = 'No procede';
+                    }
 
-                if ($chore->start_communication_status == 0) {
-                    $chore['start_communication_status_name'] = 'Pendiente';
-                } else if ($chore->start_communication_status == 1) {
-                    $chore['start_communication_status_name'] = 'Realizada';
-                } else if ($chore->start_communication_status == 2) {
-                    $chore['start_communication_status_name'] = 'No procede';
-                }
+                    if ($chore->start_communication_status == 0) {
+                        $chore['start_communication_status_name'] = 'Pendiente';
+                    } else if ($chore->start_communication_status == 1) {
+                        $chore['start_communication_status_name'] = 'Realizada';
+                    } else if ($chore->start_communication_status == 2) {
+                        $chore['start_communication_status_name'] = 'No procede';
+                    }
 
-                if ($chore->close_communication_status == 0) {
-                    $chore['close_communication_status_name'] = 'Pendiente';
-                } else if ($chore->close_communication_status == 1) {
-                    $chore['close_communication_status_name'] = 'Realizada';
-                } else if ($chore->close_communication_status == 2) {
-                    $chore['close_communication_status_name'] = 'No procede';
-                }
+                    if ($chore->close_communication_status == 0) {
+                        $chore['close_communication_status_name'] = 'Pendiente';
+                    } else if ($chore->close_communication_status == 1) {
+                        $chore['close_communication_status_name'] = 'Realizada';
+                    } else if ($chore->close_communication_status == 2) {
+                        $chore['close_communication_status_name'] = 'No procede';
+                    }
 
-                if ($chore->invoiced_status == 0) {
-                    $chore['invoiced_status_name'] = 'Pendiente';
-                } else if ($chore->invoiced_status == 1) {
-                    $chore['invoiced_status_name'] = 'Realizada';
-                } else if ($chore->invoiced_status == 2) {
-                    $chore['invoiced_status_name'] = 'No procede';
-                }
+                    if ($chore->invoiced_status == 0) {
+                        $chore['invoiced_status_name'] = 'Pendiente';
+                    } else if ($chore->invoiced_status == 1) {
+                        $chore['invoiced_status_name'] = 'Realizada';
+                    } else if ($chore->invoiced_status == 2) {
+                        $chore['invoiced_status_name'] = 'No procede';
+                    }
 
-                if ($chore->bonus_sent_status == 0) {
-                    $chore['bonus_sent_status_name'] = 'Pendiente';
-                } else if ($chore->bonus_sent_status == 1) {
-                    $chore['bonus_sent_status_name'] = 'Realizada';
-                } else if ($chore->bonus_sent_status == 2) {
-                    $chore['bonus_sent_status_name'] = 'No procede';
+                    if ($chore->bonus_sent_status == 0) {
+                        $chore['bonus_sent_status_name'] = 'Pendiente';
+                    } else if ($chore->bonus_sent_status == 1) {
+                        $chore['bonus_sent_status_name'] = 'Realizada';
+                    } else if ($chore->bonus_sent_status == 2) {
+                        $chore['bonus_sent_status_name'] = 'No procede';
+                    }
+                    return response()->json([
+                        'status' => 200,
+                        'chore' => $chore
+                    ]);
                 }
                 return response()->json([
-                    'status' => 200,
-                    'chore' => $chore
+                    'status' => 404,
+                    'message' => 'Tarea no encontrada'
                 ]);
             }
         } catch (\Exception $e){
@@ -391,9 +391,21 @@ class ChoreController extends BaseController
      * @param $id
      * @return \Illuminate\Http\JsonResponse|void
      */
-    public function destroy($id){
+    public function destroy($id, Request $request){
         if ($id) {
             try {
+               $mainCompanyId = GeneralHelpers::urlObtainCompanyId($request->headers->get('origin'), Auth::id());
+
+                $chore = Chore::where('id', $id)
+                    ->FilterMainCompany($mainCompanyId)
+                    ->first();
+
+                if (!$chore) {
+                    return response()->json([
+                        'status' => 404,
+                        'message' => 'Tarea no encontrada'
+                    ]);
+                }
                 Chore::destroy($id);
                 return response()->json([
                     'status' => 200
